[FASE 6] Separar SportMonksAPI en clases especializadas (facade pattern)
- SportMonksApiBase: clase abstracta con execute(), gson(), withIncludes(), localeUrl()
- RateLimitTracker: estado compartido de rate limit entre sub-APIs
- FootballApi, OddsApi, PredictionsApi, CoreApi: instanciables independientemente
- SportMonksAPI: facade con delegación completa (backwards compatible)
- 282 tests passing
