+++ src/main/java/es/com/kete1987/sportmonks/library/RateLimitTracker.java
@@ -1,0 +1,6 @@
+package es.com.kete1987.sportmonks.library;
+
+class RateLimitTracker {
+    volatile String total;
+    volatile String remaining;
+}
